server/requirements.txt
Werkzeug==0.12.2
Flask
requests==2.18.4
opencc-python-reimplemented
chardet
Flask-Sockets
gevent==1.3.4
greenlet==0.4.13
